+++ Rubrik/Private/Get-RubrikAPIData.ps1
@@ -788,7 +788,7 @@ function Get-RubrikAPIData($endpoint) {
         }
         'Get-RubrikProxySetting'           = @{
             '1.0' = @{
-                Description = 'Returns details of the configured proxy server within the Rubrik cluster'
+                Description = 'Get Rubrik Node Proxy Configuration'
                 URI         = '/api/internal/node_management/proxy_config'
                 Method      = 'Get'
                 Body        = ''
@@ -796,6 +796,7 @@ function Get-RubrikAPIData($endpoint) {
                 Result      = ''
                 Filter      = ''
                 Success     = '200'
+                ObjectTName = 'Rubrik.ProxySetting'
             }
         }
         'Get-RubrikReport'             = @{
@@ -1910,6 +1911,18 @@ function Get-RubrikAPIData($endpoint) {
                 Success     = '202'
             }
         }
+        'Remove-RubrikProxySetting'              = @{
+            '1.0' = @{
+                Description = 'Remove Rubrik Node Proxy Configuration'
+                URI         = '/api/internal/node_management/proxy_config'
+                Method      = 'Delete'
+                Body        = ''
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '204'
+            }
+        }
         'Remove-RubrikOrganization'           = @{
             '1.0' = @{
                 Description = 'Remove an organization from a Rubrik cluster'
@@ -2186,6 +2199,25 @@ function Get-RubrikAPIData($endpoint) {
                 Success     = '200'
             }
         }
+        'Set-RubrikProxySetting'              = @{
+            '1.0' = @{
+                Description = 'Remove Rubrik Node Proxy Configuration'
+                URI         = '/api/internal/node_management/proxy_config'
+                Method      = 'Patch'
+                Body        = @{
+                    host     = 'host'
+                    port     = 'port'
+                    protocol = 'protocol'
+                    username = 'username'
+                    password = 'password'
+                }
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '200'
+                ObjectTName = 'Rubrik.ProxySetting'
+            }
+        }
         'Set-RubrikNutanixVM'                 = @{
             '1.0' = @{
                 Description = 'Update VM with specified properties'
@@ -2194,7 +2226,6 @@ function Get-RubrikAPIData($endpoint) {
                 Body        = @{
                     snapshotConsistencyMandate = 'snapshotConsistencyMandate'
                     isPaused                 = 'isPaused'
-
                 }
                 Query       = ''
                 Result      = ''