+++ Tests/Set-RubrikProxySetting.Tests.ps1
@@ -5,7 +5,7 @@ foreach ( $privateFunctionFilePath in ( Get-ChildItem -Path './Rubrik/Private' |
     . $privateFunctionFilePath
 }
 
-Describe -Name 'Public/Get-RubrikProxySetting' -Tag 'Public', 'Get-RubrikProxySetting' -Fixture {
+Describe -Name 'Public/Set-RubrikProxySetting' -Tag 'Public', 'Set-RubrikProxySetting' -Fixture {
     #region init
     $global:rubrikConnection = @{
         id      = 'test-id'
@@ -22,28 +22,50 @@ Describe -Name 'Public/Get-RubrikProxySetting' -Tag 'Public', 'Get-RubrikProxySe
     Context -Name 'Returned Results' {
         Mock -CommandName Test-RubrikConnection -Verifiable -ModuleName 'Rubrik' -MockWith { }
         Mock -CommandName Submit-Request -Verifiable -ModuleName 'Rubrik' -MockWith {
-            [PSCustomObject]@{ 
+            [pscustomobject]@{ 
                 'host'     = 'proxy.server.com' 
                 'protocol' = 'https'
             }
         }
-        It -Name 'No parameters returns all results' -Test {
-            @( Get-RubrikProxySetting | Measure-Object).Count |
+        It -Name 'Set Rubrik Proxy Settings returns an object' -Test {
+            @( Set-RubrikProxySetting -proxyhostname willreturnone -protocol HTTP | Measure-Object).Count |
                 Should -BeExactly 1
         }
         It -Name 'Host is proxy.server.com' -Test {
-            ( Get-RubrikProxySetting ).host |
+            ( Set-RubrikProxySetting -proxyhostname 'proxy.server.com' -protocol SOCKS5).host |
                 Should -BeExactly 'proxy.server.com'
-        }
-        
+        }        
         It -Name 'NodeIPAddress property is properly populated' -Test {
-            @( Get-RubrikProxySetting -IPAddress 10.10.10.10).nodeipaddress |
+            (Set-RubrikProxySetting -IPAddress 10.10.10.10 -Host build.rubrik.com -Port 8080 -Protocol HTTPS).nodeipaddress |
                 Should -BeExactly '10.10.10.10'
         }
 
         Assert-VerifiableMock
         Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Exactly 3
         Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Exactly 3
+    }
+    
+    Context -Name 'Validate that protocol is case insensitive' {
+        Mock -CommandName Test-RubrikConnection -Verifiable -ModuleName 'Rubrik' -MockWith { }
+        Mock -CommandName Submit-Request -Verifiable -ModuleName 'Rubrik' -MockWith { }
+        
+        It -Name 'Verify Lowercase changed to upper case' -Test {
+            $Output = & {
+                Set-RubrikProxySetting -proxyhostname 'proxy.server.com' -protocol socks5 -Verbose 4>&1
+            }
+            (-join $Output) | Should -BeLikeExactly '*protocol*SOCKS5*'
+        }
+        
+        It -Name 'Verify mixed case is changed to upper case' -Test {
+            $Output = & {
+                Set-RubrikProxySetting -proxyhostname 'proxy.server.com' -protocol HtTpS -Verbose 4>&1
+            }
+            (-join $Output) | Should -BeLikeExactly '*protocol*HTTPS*'
+        }
+        
+        Assert-VerifiableMock
+        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Exactly 2
+        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik' -Exactly 2
     }
     
     Context -Name 'Function should return correct object type' {
@@ -56,7 +78,7 @@ Describe -Name 'Public/Get-RubrikProxySetting' -Tag 'Public', 'Get-RubrikProxySe
         }
         
         It -Name 'Should have a custom object type defined' -Test {
-            (Get-RubrikProxySetting).psobject.typenames |
+            (Set-RubrikProxySetting -proxyhostname 'proxy.server.com' -protocol SOCKS5).psobject.typenames |
                 Should -Contain 'Rubrik.ProxySetting'
         }